refactor: move API into service and dumb down component
This introduces a couple of changes, mostly to decouple the API service
from the component so we can control the data shape and also remove
any data we don't care about until we need it.

The App now controls all state until we need to move it down to a
container component and the PokemonList can be just what it is used
for so we can visually change it without having to worry about its API.

This is totally arbitrary but I also lower camel-cased all the filenames
That is totally bike-shedding and I just think it looks nicer to have
the folder/file names sort of disappear.